re-arranged the model folder
